[SPARK-38711][PYTHON][SS] Refactor pyspark.sql.streaming module
### What changes were proposed in this pull request?

This PR changes `pyspark/sql/streaming.py` module to a package with two files: `pyspark/sql/streaming/query.py` and `pyspark/sql/streaming/readwriter.py`. This is similar with `pyspark/sql/dataframe.py` and `pyspark/sql/readwriter.py`.

There should be no user facing change because I kept the existing import at `pyspark/sql/streaming/__init__.py`.

### Why are the changes needed?

To make the codes easier to read.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Existing test cases should cover this change.

Closes #36023 from HyukjinKwon/refactoring-streaming-module.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
